requirements.txt
aiohttp
pexpect
requests
systemd_python
paramiko
pyserial
setuptools
hidapi
pysnmp
pyasn1
pyusb
